:truck: style: Move componente de learning
deixa todos os componentes na pasta components
