+++ src/components/LearningCard/style.css
